[Refactor] UI - Policies: Extract PipelineTestPanel from pipeline_flow_builder
Extract the PipelineTestPanel component (~420 lines) and shared utilities
from the 1583-line pipeline_flow_builder.tsx into separate files.

New files:
- pipeline_icons.tsx: icons, constants, and utility functions
- pipeline_test_drawer.tsx: PipelineTestPanel component
- pipeline_icons.test.ts: 19 tests for utility functions
- pipeline_test_drawer.test.tsx: 11 tests for PipelineTestPanel
