+++ ui/litellm-dashboard/src/components/policies/pipeline_icons.tsx
@@ -1,0 +1,200 @@
+import React from "react";
+import { PlusIcon } from "@heroicons/react/outline";
+import { PipelineStep, GuardrailPipeline } from "./types";
+import type { CompliancePrompt } from "../../data/compliancePrompts";
+import { getComplianceDatasetPrompts, getFrameworks } from "../../data/compliancePrompts";
+import { Policy } from "./types";
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Constants
+// ─────────────────────────────────────────────────────────────────────────────
+
+export const TEST_SOURCE_QUICK = "quick_chat";
+export const TEST_SOURCE_ALL = "__all__";
+
+export const ACTION_OPTIONS = [
+  { label: "Next Step", value: "next" },
+  { label: "Allow", value: "allow" },
+  { label: "Block", value: "block" },
+  { label: "Custom Response", value: "modify_response" },
+];
+
+export const ACTION_LABELS: Record<string, string> = {
+  allow: "Allow",
+  block: "Block",
+  next: "Next Step",
+  modify_response: "Custom Response",
+};
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Utility functions
+// ─────────────────────────────────────────────────────────────────────────────
+
+export function getPromptsForTestSource(source: string): CompliancePrompt[] {
+  if (source === TEST_SOURCE_QUICK) return [];
+  if (source === TEST_SOURCE_ALL) return getComplianceDatasetPrompts();
+  const fw = getFrameworks().find((f) => f.name === source);
+  return fw ? fw.categories.flatMap((c) => c.prompts) : [];
+}
+
+export function createDefaultStep(): PipelineStep {
+  return {
+    guardrail: "",
+    on_pass: "next",
+    on_fail: "block",
+    pass_data: false,
+    modify_response_message: null,
+  };
+}
+
+export function insertStep(steps: PipelineStep[], atIndex: number): PipelineStep[] {
+  const newSteps = [...steps];
+  newSteps.splice(atIndex, 0, createDefaultStep());
+  return newSteps;
+}
+
+export function removeStep(steps: PipelineStep[], index: number): PipelineStep[] {
+  if (steps.length <= 1) return steps;
+  const newSteps = [...steps];
+  newSteps.splice(index, 1);
+  return newSteps;
+}
+
+export function updateStepAtIndex(
+  steps: PipelineStep[],
+  index: number,
+  updated: Partial<PipelineStep>
+): PipelineStep[] {
+  return steps.map((s, i) => (i === index ? { ...s, ...updated } : s));
+}
+
+/**
+ * Derives a pipeline from a policy. When the policy has a pipeline, use it.
+ * When it only has guardrails_add (legacy/simple form), convert those guardrails
+ * into pipeline steps in order.
+ */
+export function derivePipelineFromPolicy(policy: Policy | null | undefined): GuardrailPipeline {
+  if (!policy) {
+    return { mode: "pre_call", steps: [createDefaultStep()] };
+  }
+  if (policy.pipeline?.steps?.length) {
+    return policy.pipeline;
+  }
+  const guardrails = policy.guardrails_add || [];
+  if (guardrails.length > 0) {
+    return {
+      mode: policy.pipeline?.mode ?? "pre_call",
+      steps: guardrails.map((g) => ({
+        guardrail: g,
+        on_pass: "next" as const,
+        on_fail: "block" as const,
+        pass_data: false,
+        modify_response_message: null,
+      })),
+    };
+  }
+  return { mode: "pre_call", steps: [createDefaultStep()] };
+}
+
+export function complianceMatchExpected(expected: "pass" | "fail", terminalAction: string): boolean {
+  if (expected === "pass") {
+    return terminalAction === "allow" || terminalAction === "modify_response";
+  }
+  return terminalAction === "block";
+}
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Icons
+// ─────────────────────────────────────────────────────────────────────────────
+
+export const GuardrailIcon: React.FC = () => (
+  <div
+    style={{
+      width: 28,
+      height: 28,
+      borderRadius: "50%",
+      backgroundColor: "#eef2ff",
+      display: "flex",
+      alignItems: "center",
+      justifyContent: "center",
+      flexShrink: 0,
+    }}
+  >
+    <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="#6366f1" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <circle cx="12" cy="12" r="10" />
+      <path d="M12 8v4" />
+    </svg>
+  </div>
+);
+
+export const PlayIcon: React.FC = () => (
+  <div
+    style={{
+      width: 28,
+      height: 28,
+      borderRadius: "50%",
+      backgroundColor: "#f3f4f6",
+      display: "flex",
+      alignItems: "center",
+      justifyContent: "center",
+      flexShrink: 0,
+    }}
+  >
+    <svg width="12" height="12" viewBox="0 0 24 24" fill="#6b7280" stroke="none">
+      <polygon points="6,3 20,12 6,21" />
+    </svg>
+  </div>
+);
+
+export const PassIcon: React.FC = () => (
+  <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="#22c55e" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" style={{ flexShrink: 0 }}>
+    <circle cx="12" cy="12" r="10" />
+    <path d="M9 12l2 2 4-4" />
+  </svg>
+);
+
+export const FailIcon: React.FC = () => (
+  <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="#f87171" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" style={{ flexShrink: 0 }}>
+    <circle cx="12" cy="12" r="10" />
+  </svg>
+);
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Connector
+// ─────────────────────────────────────────────────────────────────────────────
+
+interface ConnectorProps {
+  onInsert: () => void;
+}
+
+export const Connector: React.FC<ConnectorProps> = ({ onInsert }) => (
+  <div className="flex flex-col items-center" style={{ height: 56 }}>
+    <div style={{ width: 1, flex: 1, backgroundColor: "#d1d5db" }} />
+    <button
+      onClick={onInsert}
+      className="flex items-center justify-center"
+      style={{
+        width: 24,
+        height: 24,
+        borderRadius: "50%",
+        border: "1px solid #d1d5db",
+        backgroundColor: "#fff",
+        cursor: "pointer",
+        zIndex: 1,
+        transition: "all 0.15s ease",
+      }}
+      onMouseEnter={(e) => {
+        e.currentTarget.style.borderColor = "#6366f1";
+        e.currentTarget.style.backgroundColor = "#eef2ff";
+      }}
+      onMouseLeave={(e) => {
+        e.currentTarget.style.borderColor = "#d1d5db";
+        e.currentTarget.style.backgroundColor = "#fff";
+      }}
+      title="Insert step"
+    >
+      <PlusIcon style={{ width: 12, height: 12, color: "#9ca3af" }} />
+    </button>
+    <div style={{ width: 1, flex: 1, backgroundColor: "#d1d5db" }} />
+  </div>
+);
